+++ generator_tests/test/doc/tiny_test.dart
@@ -171,7 +171,7 @@ class _TinyFormBuilderState extends State<TinyFormBuilder> {
   }
 }
 
-class TinyForm {
+class TinyForm implements FormModel<Tiny> {
   TinyForm(this.tiny, this.form, this.path) {}
 
   static String emailControlName = "email";