+++ generator_tests/test/freezed/freezed_test.dart
@@ -171,7 +171,7 @@ class _TestFormBuilderState extends State<TestFormBuilder> {
   }
 }
 
-class TestForm {
+class TestForm implements FormModel<Test> {
   TestForm(this.test, this.form, this.path) {}
 
   static String genderControlName = "gender";