+++ generator_tests/test/group_nullable_test.dart
@@ -220,7 +220,7 @@ class _GroupFormBuilderState extends State<GroupFormBuilder> {
   }
 }
 
-class GroupForm {
+class GroupForm implements FormModel<Group> {
   GroupForm(this.group, this.form, this.path) {
     personalForm = PersonalForm(group.personal, form, pathBuilder('personal'));
     phoneForm = PhoneForm(group.phone, form, pathBuilder('phone'));
@@ -411,7 +411,7 @@ class GroupForm {
           disabled: false);
 }
 
-class PersonalForm {
+class PersonalForm implements FormModel<Personal> {
   PersonalForm(this.personal, this.form, this.path) {}
 
   static String nameControlName = "name";
@@ -516,7 +516,7 @@ class PersonalForm {
           disabled: false);
 }
 
-class PhoneForm {
+class PhoneForm implements FormModel<Phone> {
   PhoneForm(this.phone, this.form, this.path) {}
 
   static String phoneNumberControlName = "phoneNumber";
@@ -621,7 +621,7 @@ class PhoneForm {
           disabled: false);
 }
 
-class AddressForm {
+class AddressForm implements FormModel<Address> {
   AddressForm(this.address, this.form, this.path) {}
 
   static String streetControlName = "street";