+++ packages/reactive_forms_annotations/lib/reactive_forms_annotations.dart
@@ -4,3 +4,4 @@ export 'src/reactive_form_annotation.dart';
 export 'src/form_group_annotation.dart';
 export 'src/form_array_annotation.dart';
 export 'src/form_control_annotation.dart';
+export 'src/form_model.dart';