+++ packages/reactive_forms_annotations/lib/src/form_model.dart
@@ -1,0 +1,7 @@
+import 'package:reactive_forms/reactive_forms.dart';
+
+abstract class FormModel<TModel> {
+  FormModel({required this.form});
+  final FormGroup form;
+  TModel get model;
+}