+++ packages/reactive_forms_generator/example/lib/docs/array_group/delivery_list.gform.dart
@@ -129,7 +129,7 @@ class _DeliveryListFormBuilderState extends State<DeliveryListFormBuilder> {
   }
 }
 
-class DeliveryListForm {
+class DeliveryListForm implements FormModel<DeliveryList> {
   DeliveryListForm(this.deliveryList, this.form, this.path) {
     deliveryListDeliveryPointForm = deliveryList.deliveryList
         .asMap()
@@ -244,7 +244,7 @@ class DeliveryListForm {
           disabled: false);
 }
 
-class DeliveryPointForm {
+class DeliveryPointForm implements FormModel<DeliveryPoint> {
   DeliveryPointForm(this.deliveryPoint, this.form, this.path) {
     addressForm =
         AddressForm(deliveryPoint.address, form, pathBuilder('address'));
@@ -357,7 +357,7 @@ class DeliveryPointForm {
           disabled: false);
 }
 
-class AddressForm {
+class AddressForm implements FormModel<Address> {
   AddressForm(this.address, this.form, this.path) {}
 
   static String streetControlName = "street";