+++ packages/reactive_forms_generator/example/lib/routes/login_nullable_form.dart
@@ -51,7 +51,7 @@ class LoginNullableFormWidget extends StatelessWidget {
                 ),
               ),
               const SizedBox(height: 16.0),
-              ReactiveDropdownSearch<String>(
+              ReactiveDropdownSearch<String, String>(
                 formControl: formModel.themeControl,
                 decoration: InputDecoration(
                   helperText: '',
@@ -60,7 +60,7 @@ class LoginNullableFormWidget extends StatelessWidget {
                 ),
                 mode: Mode.MENU,
                 hint: "Select a theme",
-                showSelectedItem: true,
+                // showSelectedItem: true,
                 items: [
                   "light",
                   "dark",
@@ -69,7 +69,7 @@ class LoginNullableFormWidget extends StatelessWidget {
                 showClearButton: true,
               ),
               const SizedBox(height: 16.0),
-              ReactiveDropdownSearch<UserMode>(
+              ReactiveDropdownSearch<UserMode, UserMode>(
                 formControl: formModel.modeControl,
                 decoration: InputDecoration(
                   helperText: '',
@@ -78,7 +78,7 @@ class LoginNullableFormWidget extends StatelessWidget {
                 ),
                 mode: Mode.MENU,
                 hint: "Select a theme",
-                showSelectedItem: true,
+                // showSelectedItem: true,
                 items: [
                   UserMode.admin,
                   UserMode.user,