+++ packages/reactive_forms_generator/lib/src/library_builder.dart
@@ -1,7 +1,5 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:code_builder/code_builder.dart';
-import 'package:dart_style/dart_style.dart';
-import 'package:reactive_forms_generator/src/extensions.dart';
 import 'package:reactive_forms_generator/src/form_generator.dart';
 import 'package:reactive_forms_generator/src/reactive_forms/reactive_form.dart';
 import 'package:reactive_forms_generator/src/reactive_forms/reactive_form_builder.dart';